修复城市管理界面标题栏BUG，修复返回温度数据类型不一致问题

+++ app/src/main/java/com/example/administrator/justfortest2/gson/DailyTemp.java
@@ -5,7 +5,7 @@ package com.example.administrator.justfortest2.gson;
  */
 
 public class DailyTemp {
-    public int max;
-    public int min;
+    public float max;
+    public float min;
     public String date;
 }
